backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
python-dotenv
pydantic>=2.11.7
httpx
ffmpeg-python
pydub
supabase
python-multipart


